refactor update folder
