+++ src/display/components/bar.js
@@ -5,7 +5,7 @@ import { getTexture } from '../../assets/textures/texture';
 import { validate } from '../../utils/vaildator';
 import { componentPipeline } from '../change/component-pipeline';
 import { TextureStyle } from '../data-schema/component-schema';
-import { updateObject } from '../update-object';
+import { updateObject } from '../update/update-object';
 
 const barSchema = z.object({
   texture: z.union([z.string(), TextureStyle]),
