+++ src/display/elements/item.js
@@ -1,5 +1,5 @@
 import { elementPipeline } from '../change/element-pipeline';
-import { updateObject } from '../update-object';
+import { updateObject } from '../update/update-object';
 import { createContainer } from '../utils';
 
 export const createItem = (config) => {
