+++ src/display/update/update-object.js
@@ -1,5 +1,5 @@
-import { deepMerge } from '../utils/deepmerge/deepmerge';
-import { changeProperty } from './change/change';
+import { deepMerge } from '../../utils/deepmerge/deepmerge';
+import { changeProperty } from '../change/change';
 
 const DEFAULT_EXCEPTION_KEYS = new Set(['position']);
