+++ src/patchMap.js
@@ -2,7 +2,7 @@ import { Application, Assets } from 'pixi.js';
 import { isValidationError } from 'zod-validation-error';
 import { undoRedoManager } from './command';
 import { draw } from './display/draw';
-import { update } from './display/update';
+import { update } from './display/update/update';
 import { dragSelect } from './events/drag-select';
 import { select } from './events/single-select';
 import {
